Move fetch logic from Chat to ConversationListContainer
